fix: refactor and subscribe windows agent to rustdesk
