+++ internal/agent/rustdesk/rustdesk_linux.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 package rustdesk
 
 import (
